warning culling

+++ fork/os/process.c
@@ -1187,11 +1187,11 @@ The items that need to be fixed for this lab in DLXOS are:
   dbprintf('m', "ProcessRealFork(%d): Calling copying system stack contents.\n", GetCurrentPid());
   bcopy((char *)currentPCB->currentSavedFrame[PROCESS_STACK_PTBASE], (char *)newSystemStackPage, MEM_PAGESIZE);
 
-  childPCB->currentSavedFrame = (childPCB->currentSavedFrame && 0x00000FFF) | newSystemStackPage;
+  childPCB->currentSavedFrame = (uint32)((childPCB->currentSavedFrame && 0x00000FFF) | newSystemStackPage);
   dbprintf('m', "ProcessRealFork(%d): updated childPCB currentSavedFrame to 0x%x.\n", GetCurrentPid(), (uint32)childPCB->currentSavedFrame);
   childPCB->currentSavedFrame[PROCESS_STACK_PTBASE] = (childPCB->currentSavedFrame[PROCESS_STACK_PTBASE] && 0x00000FFF) | newSystemStackPage;
   dbprintf('m', "ProcessRealFork(%d): updated childPCB PTBase to 0x%x.\n", GetCurrentPid(), (uint32)childPCB->currentSavedFrame[PROCESS_STACK_PTBASE]);
-  childPCB->sysStackPtr = (childPCB->sysStackPtr && 0x00000FFF) | newSystemStackPage;
+  childPCB->sysStackPtr = (uint32)((childPCB->sysStackPtr && 0x00000FFF) | newSystemStackPage);
   dbprintf('m', "ProcessRealFork(%d): updated childPCB sysStackPtr to 0x%x.\n", GetCurrentPid(), (uint32)childPCB->sysStackPtr);
 
   for (i = 0; i < MEM_NUM_PAGE_TABLE_ENTRIES; i++)
